--- src/main/kotlin/com/vibetrip/vibetripserver/album/presentation/dto/AlbumCreateResponse.kt
@@ -1,5 +1,0 @@
-package com.vibetrip.vibetripserver.album.presentation.dto
-
-data class AlbumCreateResponse(
-    val albumId: Long,
-)
