+++ src/main/kotlin/com/vibetrip/vibetripserver/album/presentation/dto/response/AlbumCreateResponse.kt
@@ -1,0 +1,5 @@
+package com.vibetrip.vibetripserver.album.presentation.dto.response
+
+data class AlbumCreateResponse(
+    val albumId: Long,
+)
